implementation of __array__ and __array_ufunc__ on ContainerOperand
